Includes protocol and Process* scripts
Protocol is a work in progress, and is not meant to replace the detailed
protocol for Dynein experiments that was, at one point, found on the
Yildiz lab wiki.

ProcessAll.m automatically works up all the spotlists in the current
directory.

ProcessDyns.py is a script for calling CombineWHSpots, and assumes it is
being run from within IPython with a -i flag.  You must adjust the
MAP variable, which points to the Holes map, and the OFFSETS variable,
which is a dictionary, where keys are the offset map number, and values
are the dynein movie number.  It's... fragile, at best.